feat(auth-service): show "Powered by Certified" footer on handle picker and recovery pages
Extracts the footer HTML + SVG read into a shared POWERED_BY_HTML /
POWERED_BY_CSS pair in lib/page-helpers.ts so the SVG is read once,
then applies it to /auth/choose-handle and both /auth/recover steps.
account-login is refactored to consume the same helper.
