+++ .changeset/handle-recovery-powered-by-footer.md
@@ -1,0 +1,9 @@
+---
+'ePDS': patch
+---
+
+"Powered by Certified" footer now appears on the handle picker and account recovery pages.
+
+**Affects:** End users
+
+**End users:** the "Choose your handle" page (shown to new users after email verification) and both Account Recovery steps (backup-email entry and recovery-code entry) now display the same "Powered by Certified" footer that the main sign-in and Account Settings sign-in flows already show, so the branding is consistent across every auth-service page.
